refactor(workspace): reorganize benchmarks into __benchmarks__/ directory
Split the 1,302-line benchmark.test.ts mega-file and two scattered
comparison files into 7 focused benchmark files, each answering one
question. Shared fixtures extracted to helpers.ts.

Files moved:
- workspace/benchmark.test.ts → split into 6 files
- y-keyvalue/ymap-simplicity-case.test.ts → ykv-vs-ymap.bench.test.ts
- y-keyvalue/y-keyvalue-comparison.test.ts → conflict-resolution.bench.test.ts

Also fixes pre-existing createEncryptedYkvLww(yarray, {}) calls that
broke when the API changed to accept no second argument.
