+++ packages/workspace/src/__benchmarks__/conflict-resolution.bench.test.ts
@@ -1,18 +1,23 @@
 /**
- * YKeyValue vs YKeyValueLww - Comparison Tests
+ * Conflict Resolution Comparison: YKeyValue vs YKeyValueLww
  *
- * These tests compare positional conflict resolution and timestamp-based LWW
- * conflict resolution under the same concurrent edit scenarios. They also measure
- * storage overhead and confirm API parity for migration safety.
+ * Answers: "How do the two YKeyValue implementations differ in conflict behavior?"
  *
- * Key behaviors:
- * - Equal scenarios produce different winners depending on conflict strategy.
- * - LWW adds predictable storage overhead while preserving the same public API.
+ * Compares positional conflict resolution (clientID-based, unpredictable winner)
+ * against timestamp-based LWW (later edit always wins) under identical concurrent
+ * edit scenarios. Also measures storage overhead and confirms API parity.
  */
+
 import { describe, expect, test } from 'bun:test';
 import * as Y from 'yjs';
-import { YKeyValue, type YKeyValueEntry } from './y-keyvalue';
-import { YKeyValueLww, type YKeyValueLwwEntry } from './y-keyvalue-lww';
+import {
+	YKeyValue,
+	type YKeyValueEntry,
+} from '../shared/y-keyvalue/y-keyvalue.js';
+import {
+	YKeyValueLww,
+	type YKeyValueLwwEntry,
+} from '../shared/y-keyvalue/y-keyvalue-lww.js';
 
 /** Helper to create a YKeyValue with fresh doc */
 function createKv<T>() {
@@ -28,7 +33,11 @@ function createKvLww<T>() {
 	return { doc, array, kv: new YKeyValueLww(array) };
 }
 
-describe('Conflict Resolution: YKeyValue vs YKeyValueLww', () => {
+// ═══════════════════════════════════════════════════════════════════════════════
+// Conflict Resolution
+// ═══════════════════════════════════════════════════════════════════════════════
+
+describe('conflict resolution: YKeyValue vs YKeyValueLww', () => {
 	/**
 	 * THE KEY DIFFERENCE: Same concurrent edits, different winners.
 	 *
@@ -144,7 +153,11 @@ describe('Conflict Resolution: YKeyValue vs YKeyValueLww', () => {
 	});
 });
 
-describe('Storage Comparison', () => {
+// ═══════════════════════════════════════════════════════════════════════════════
+// Storage Comparison
+// ═══════════════════════════════════════════════════════════════════════════════
+
+describe('storage comparison', () => {
 	test('LWW entries serialize larger than positional entries', () => {
 		const { array: arr } = createKv<string>();
 		const { array: arrLww } = createKvLww<string>();
@@ -196,7 +209,11 @@ describe('Storage Comparison', () => {
 	});
 });
 
-describe('API Compatibility', () => {
+// ═══════════════════════════════════════════════════════════════════════════════
+// API Compatibility
+// ═══════════════════════════════════════════════════════════════════════════════
+
+describe('API compatibility', () => {
 	test('both implementations have identical API surface', () => {
 		const { kv: kv1 } = createKv<string>();
 		const { kv: kv2 } = createKvLww<string>();
