+++ packages/workspace/src/__benchmarks__/storage-overhead.bench.test.ts
@@ -1,0 +1,299 @@
+/**
+ * Storage Overhead Benchmarks
+ *
+ * Answers: "How big is my data in Y.Doc vs raw JSON?"
+ *
+ * Measures CRDT overhead ratios, per-row byte costs, size growth after updates,
+ * and practical ceiling estimates for row counts.
+ */
+
+import { describe, test } from 'bun:test';
+import * as Y from 'yjs';
+import { createTables } from '../workspace/create-tables.js';
+import {
+	formatBytes,
+	generateId,
+	heavyNoteDefinition,
+	makeHeavyRow,
+	noteDefinition,
+	postDefinition,
+} from './helpers.js';
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// CRDT Overhead vs Raw JSON
+// ═══════════════════════════════════════════════════════════════════════════════
+
+describe('CRDT overhead vs raw JSON', () => {
+	test('small row: actual payload vs Y.Doc overhead', () => {
+		const ydoc = new Y.Doc();
+		const tables = createTables(ydoc, { posts: postDefinition });
+
+		// Example row
+		const row = { id: 'id-000001', title: 'Post 1', views: 42 };
+		const jsonPayload = JSON.stringify(row);
+		console.log('\n=== SMALL ROW ANALYSIS ===');
+		console.log(`Row data: ${jsonPayload}`);
+		console.log(`JSON payload size: ${jsonPayload.length} bytes`);
+
+		// What Y.js actually stores (from YKeyValueLww):
+		// { key: 'id-000001', val: { id: 'id-000001', ... }, ts: 1706200000000 }
+		const yEntry = { key: row.id, val: row, ts: Date.now() };
+		console.log(`Y.js wrapper JSON: ${JSON.stringify(yEntry).length} bytes`);
+		console.log(
+			`Overhead: +${JSON.stringify(yEntry).length - jsonPayload.length} bytes (ID stored twice + timestamp)`,
+		);
+
+		// Insert 1000 rows
+		for (let i = 0; i < 1_000; i++) {
+			tables.posts.set({
+				id: generateId(i),
+				title: `Post ${i}`,
+				views: i,
+				_v: 1,
+			});
+		}
+
+		const encoded = Y.encodeStateAsUpdate(ydoc);
+		const pureJsonSize = jsonPayload.length * 1_000;
+		console.log(`\nWith 1,000 rows:`);
+		console.log(`  Y.Doc binary: ${(encoded.byteLength / 1024).toFixed(2)} KB`);
+		console.log(`  Per row: ${(encoded.byteLength / 1_000).toFixed(0)} bytes`);
+		console.log(
+			`  Pure JSON would be: ~${(pureJsonSize / 1024).toFixed(0)} KB`,
+		);
+		console.log(
+			`  CRDT overhead: ${((encoded.byteLength / pureJsonSize - 1) * 100).toFixed(0)}%`,
+		);
+	});
+
+	test('realistic row: notes with 500 chars of content', () => {
+		const ydoc = new Y.Doc();
+		const tables = createTables(ydoc, { notes: noteDefinition });
+
+		const sampleContent = `This is a realistic note with actual content. 
+It might contain multiple paragraphs and various formatting.
+Users typically write notes that are a few hundred characters long.
+Some notes are longer, some are shorter, but this is a reasonable average.
+Let's add a bit more to make it realistic. The quick brown fox jumps over the lazy dog.`;
+
+		const row = {
+			id: generateId(0),
+			title: 'Meeting Notes - Q4 Planning',
+			content: sampleContent,
+			tags: ['work', 'meetings', 'planning'],
+			createdAt: Date.now(),
+			updatedAt: Date.now(),
+		};
+
+		const jsonPayload = JSON.stringify(row);
+		console.log('\n=== REALISTIC ROW ANALYSIS ===');
+		console.log(`Content length: ${sampleContent.length} chars`);
+		console.log(`Full row JSON: ${jsonPayload.length} bytes`);
+
+		for (let i = 0; i < 1_000; i++) {
+			tables.notes.set({
+				id: generateId(i),
+				title: `Note ${i}`,
+				content: sampleContent,
+				tags: ['tag1', 'tag2'],
+				createdAt: Date.now(),
+				updatedAt: Date.now(),
+				_v: 1,
+			});
+		}
+
+		const encoded = Y.encodeStateAsUpdate(ydoc);
+		console.log(`\nWith 1,000 notes (~500 chars each):`);
+		console.log(
+			`  Y.Doc binary: ${(encoded.byteLength / 1024).toFixed(0)} KB (${(encoded.byteLength / 1024 / 1024).toFixed(2)} MB)`,
+		);
+		console.log(`  Per row: ${(encoded.byteLength / 1_000).toFixed(0)} bytes`);
+	});
+
+	test('upper ceiling estimates', () => {
+		console.log('\n=== PRACTICAL LIMITS ===');
+		console.log(
+			'Based on benchmarks (~75 bytes/small row, ~700 bytes/note):\n',
+		);
+
+		console.log('| Rows     | Small Rows  | Notes (~500 chars) |');
+		console.log('|----------|-------------|---------------------|');
+		console.log('| 1,000    | ~75 KB      | ~700 KB             |');
+		console.log('| 10,000   | ~750 KB     | ~7 MB               |');
+		console.log('| 50,000   | ~3.7 MB     | ~35 MB              |');
+		console.log('| 100,000  | ~7.5 MB     | ~70 MB              |');
+
+		console.log('\nRecommendations:');
+		console.log('  ✓ 10K rows: Sweet spot for local-first (fast, <10MB)');
+		console.log('  ⚠ 50K rows: Still works, slower inserts (~5s)');
+		console.log('  ✗ 100K+ rows: Consider pagination/archiving');
+		console.log('  Note: Deletes are O(n) - avoid repeated bulk delete cycles');
+	});
+});
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// Encoded Size at Scale
+// ═══════════════════════════════════════════════════════════════════════════════
+
+describe('encoded size at scale', () => {
+	test('Y.Doc encoded size with 1,000 / 10,000 rows', () => {
+		const ydoc1 = new Y.Doc();
+		const tables1 = createTables(ydoc1, { posts: postDefinition });
+
+		for (let i = 0; i < 1_000; i++) {
+			tables1.posts.set({
+				id: generateId(i),
+				title: `Post ${i}`,
+				views: i,
+				_v: 1,
+			});
+		}
+		const size1k = Y.encodeStateAsUpdate(ydoc1).byteLength;
+
+		const ydoc2 = new Y.Doc();
+		const tables2 = createTables(ydoc2, { posts: postDefinition });
+
+		for (let i = 0; i < 10_000; i++) {
+			tables2.posts.set({
+				id: generateId(i),
+				title: `Post ${i}`,
+				views: i,
+				_v: 1,
+			});
+		}
+		const size10k = Y.encodeStateAsUpdate(ydoc2).byteLength;
+
+		console.log(`Y.Doc size with 1,000 rows: ${(size1k / 1024).toFixed(2)} KB`);
+		console.log(
+			`Y.Doc size with 10,000 rows: ${(size10k / 1024).toFixed(2)} KB`,
+		);
+		console.log(`Bytes per row (1k): ${(size1k / 1_000).toFixed(2)}`);
+		console.log(`Bytes per row (10k): ${(size10k / 10_000).toFixed(2)}`);
+	});
+
+	test('Y.Doc size growth after updates (same rows updated 5 times)', () => {
+		const ydoc = new Y.Doc();
+		const tables = createTables(ydoc, { posts: postDefinition });
+
+		for (let i = 0; i < 1_000; i++) {
+			tables.posts.set({
+				id: generateId(i),
+				title: `Post ${i}`,
+				views: 0,
+				_v: 1,
+			});
+		}
+		const initialSize = Y.encodeStateAsUpdate(ydoc).byteLength;
+
+		for (let update = 1; update <= 5; update++) {
+			for (let i = 0; i < 1_000; i++) {
+				tables.posts.set({
+					id: generateId(i),
+					title: `Post ${i} v${update}`,
+					views: update,
+					_v: 1,
+				});
+			}
+		}
+		const finalSize = Y.encodeStateAsUpdate(ydoc).byteLength;
+
+		console.log(
+			`Initial size (1,000 rows): ${(initialSize / 1024).toFixed(2)} KB`,
+		);
+		console.log(
+			`Final size (after 5 updates each): ${(finalSize / 1024).toFixed(2)} KB`,
+		);
+		console.log(`Growth factor: ${(finalSize / initialSize).toFixed(2)}x`);
+	});
+});
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// Heavy Text: Baseline Sizes
+// ═══════════════════════════════════════════════════════════════════════════════
+
+describe('heavy text baseline sizes', () => {
+	test('5 rows with 10K chars each: baseline size', () => {
+		const ydoc = new Y.Doc();
+		const tables = createTables(ydoc, { notes: heavyNoteDefinition });
+
+		const rows = Array.from({ length: 5 }, (_, i) =>
+			makeHeavyRow(`doc-${i}`, 10_000),
+		);
+		for (const row of rows) tables.notes.set(row);
+
+		const encoded = Y.encodeStateAsUpdate(ydoc);
+		const jsonSize = rows.reduce((s, r) => s + JSON.stringify(r).length, 0);
+
+		console.log('\n=== 5 ROWS × 10K CHARS EACH ===');
+		console.log(`  Pure JSON size:    ${formatBytes(jsonSize)}`);
+		console.log(`  Y.Doc binary size: ${formatBytes(encoded.byteLength)}`);
+		console.log(
+			`  CRDT overhead:     ${((encoded.byteLength / jsonSize - 1) * 100).toFixed(1)}%`,
+		);
+		console.log(`  Per row:           ${formatBytes(encoded.byteLength / 5)}`);
+	});
+
+	test('5 rows with 50K chars each: baseline size', () => {
+		const ydoc = new Y.Doc();
+		const tables = createTables(ydoc, { notes: heavyNoteDefinition });
+
+		const rows = Array.from({ length: 5 }, (_, i) =>
+			makeHeavyRow(`doc-${i}`, 50_000),
+		);
+		for (const row of rows) tables.notes.set(row);
+
+		const encoded = Y.encodeStateAsUpdate(ydoc);
+		const jsonSize = rows.reduce((s, r) => s + JSON.stringify(r).length, 0);
+
+		console.log('\n=== 5 ROWS × 50K CHARS EACH ===');
+		console.log(`  Pure JSON size:    ${formatBytes(jsonSize)}`);
+		console.log(`  Y.Doc binary size: ${formatBytes(encoded.byteLength)}`);
+		console.log(
+			`  CRDT overhead:     ${((encoded.byteLength / jsonSize - 1) * 100).toFixed(1)}%`,
+		);
+		console.log(`  Per row:           ${formatBytes(encoded.byteLength / 5)}`);
+	});
+
+	test('5 rows with 100K chars each: baseline size', () => {
+		const ydoc = new Y.Doc();
+		const tables = createTables(ydoc, { notes: heavyNoteDefinition });
+
+		const rows = Array.from({ length: 5 }, (_, i) =>
+			makeHeavyRow(`doc-${i}`, 100_000),
+		);
+		for (const row of rows) tables.notes.set(row);
+
+		const encoded = Y.encodeStateAsUpdate(ydoc);
+		const jsonSize = rows.reduce((s, r) => s + JSON.stringify(r).length, 0);
+
+		console.log('\n=== 5 ROWS × 100K CHARS EACH ===');
+		console.log(`  Pure JSON size:    ${formatBytes(jsonSize)}`);
+		console.log(`  Y.Doc binary size: ${formatBytes(encoded.byteLength)}`);
+		console.log(
+			`  CRDT overhead:     ${((encoded.byteLength / jsonSize - 1) * 100).toFixed(1)}%`,
+		);
+		console.log(`  Per row:           ${formatBytes(encoded.byteLength / 5)}`);
+	});
+
+	test('raw text size scaling: how content size dominates', () => {
+		console.log('\n=== TEXT SIZE SCALING (single row) ===');
+		console.log('| Content Size | JSON     | Y.Doc    | Overhead |');
+		console.log('|-------------|----------|----------|----------|');
+
+		for (const chars of [1_000, 5_000, 10_000, 50_000, 100_000, 500_000]) {
+			const ydoc = new Y.Doc();
+			const tables = createTables(ydoc, { notes: heavyNoteDefinition });
+
+			const row = makeHeavyRow('doc-0', chars);
+			tables.notes.set(row);
+
+			const encoded = Y.encodeStateAsUpdate(ydoc).byteLength;
+			const jsonSize = JSON.stringify(row).length;
+			const overhead = ((encoded / jsonSize - 1) * 100).toFixed(1);
+
+			console.log(
+				`| ${formatBytes(chars).padEnd(11)} | ${formatBytes(jsonSize).padEnd(8)} | ${formatBytes(encoded).padEnd(8)} | ${overhead.padStart(5)}%   |`,
+			);
+		}
+	});
+});
