+++ packages/workspace/src/shared/y-keyvalue/y-keyvalue.ts
@@ -3,7 +3,7 @@
  *
  * **Not used in production.** The workspace uses `YKeyValueLww` (timestamp-based)
  * exclusively. This positional (rightmost-wins) implementation is kept for:
- * - Comparison tests (`y-keyvalue-comparison.test.ts`) that document WHY LWW was chosen
+ * - Comparison benchmarks (`__benchmarks__/conflict-resolution.bench.test.ts`) that document WHY LWW was chosen
  * - Reference implementation showing the simpler conflict resolution model
  *
  * Based on [y-utility](https://github.com/yjs/y-utility) (MIT License).
